fix(fixed): avoid UB edge case
Technically, negating S64_MIN *in* an s64 is undefined behavior because
-S64_MIN would be S64_MAX + 1, causing a signed integer overflow. In
practice, it produces S64_MIN again, but we have to avoid UB in the
kernel. Negating it in a u64 is well-defined: -S64_MIN == S64_MIN.

This revision does the negation while the value is still u64, then
casts the result of the full expression to s64.

+++ src/curves/driver/fixed.h
@@ -632,7 +632,7 @@ static inline s64 curves_fixed_divide(s64 dividend,
 	}
 
 	// Apply sign.
-	return sign ? -(s64)final_res : (s64)final_res;
+	return (s64)(sign ? -final_res : final_res);
 }
 
 #endif /* _CURVES_FIXED_H */
